requirements.txt
pyyaml==6.0
google-generativeai==0.3.2